fix(fe2): comment viewerState not loading correctly
